Serve MTA-STS policy under mta-sts.inclusion.gouv.fr
Scalingo was configured with a canonical domain to redirect all requests
to inclusion.gouv.fr with an HTTP 301. However, the MTA-STS policy must
be served directly, without redirects. From
https://www.mailhardener.com/tools/mta-sts-validator?domain=inclusion.gouv.fr

> An MTA-STS policy server is not allowed to redirect (RFC8461, section 3.3)

The Scalingo app has been configured to remove the canonical domain,
allowing the app to be served under mta-sts.inclusion.gouv.fr as well as
inclusion.gouv.fr.

In order to be able to handle different domain differently, a middleware
was introduced. It:
- redirects from inclusion.beta.gouv.fr → inclusion.gouv.fr
- filters requests to mta-sts.inclusion.gouv.fr to only serve expected
  routes
- serves other requests as usual